requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.10
django-bootstrap3==21.1
django-bootstrap4==21.1
django-crispy-forms==1.13.0
django-filter==21.1
django-tables2==2.4.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
importlib-metadata==2.1.2
psycopg2==2.9.2
PyJWT==1.7.1
pyspark==3.0.2
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
typing-extensions==4.0.1
zipp==3.6.0
